Refactor control modules from interpreter into own file
